requirements.txt
boto3==1.17.27
botocore==1.20.27
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
MarkupSafe==1.1.1
python-dateutil==2.8.1
PyYAML==5.4.1
s3transfer==0.3.4
six==1.15.0
urllib3==1.26.3
Werkzeug==1.0.1
